Allow Weikai to pass data_bit validation
Some components schema validate uart `data_bits` are set to a specific value but Weikai UARTs do not expose `data_bits` since they can't be configured. This is because ithere is a hardcoded constant of 8 bits from the silcon.

This change allows data_bits to be set on the UART bus but only to the 8 bits value the silcon supports. Since 8 bits is what most devices are looking for schema validation passes.

#6858